feat(loom): ACP Conversation reads the chat journal + live stream (5a)
For an `acp` session the Conversation surface swaps its data source from the
iris scrape (`/conversation`) to the durable chat journal: it paints the
`GET /chat` snapshot, then applies the `/chat/stream` SSE tail in place —
`block` upserts by (turn, seq), `delta` streams into a shadow message/thought,
`tool` tracks live tool state, `turn` drives the working indicator. The
composer posts to `/prompt` (a 202 with `queued:true` shows a "queued for next
turn" note against an optimistic user block); Stop posts `/interrupt`
(`session/cancel`). Permission cards answer via `/permissions/{id}`.

SessionConversation becomes a thin dispatcher on `session.protocol`: the
terminal-backend path moves verbatim to TerminalConversation (untouched), and
the new AcpConversation owns the journal + SSE plumbing. types.ts / api.ts mirror
the new serde shapes and endpoints by hand.
